Bugfix worksheet, formula and range parsing
The problem: Parsing a worksheet with a formula that uses range addresses eg; Sheet1!A1:D1 where C1 was an empty cell would omit C1 as a cell in Model.cells as it's an empty cell.

The fix: As formulas are parsed in ModelCompiler.build_ranges() we now ensure these empty cells are included in  Model.cells.

removed skip for tests in CONCAT, COUNT and COUNTA

Bugfix the function COUNTA

Co-Authored-By: Bradley van Ree <bradbase@users.noreply.github.com>