Add Case and parametrize_cases to tests/testing.py
The forthcoming commits will include a lot of parametrized tests, and
these two bits of code make them a lot easier to write. They may also
end up being useful elsewhere in the project. This is essentially just
vendoring my package "pytest-parametrize-cases", with a minor change to
backport it to Python 3.7 (positional-only arguments are not supported
in 3.7).