Implement "Engineering" base-conversion functions
This is about as clean as I could make the code while still hitting all
of the edge cases identified through testing. I couldn't figure out a
way to avoid the somewhat ugly special-casing that occurs when one of
the bases is decimal. There are also more type conversions than I would
like.