Add property-based tests of engineering functions
We can combine the "xlwings" and "hypothesis" libraries to create a set
of "oracle" tests. First, we define in general terms what kind of input
values we want (e.g. numbers within some range). We also specify which
Python implementation of an Excel formula is under test. We then launch
a live Excel instance and populate a cell (say, B1) such that it
contains the Excel version of the formula we want, e.g. "=DEC2BIN(A1)",
where A1 is initially blank. Then, for each value x generated by
hypothesis, we set A1=x and observe the value of B1. We also compute the
result of func(x) where func is the Python implementation of the Excel
function. If B1 != func(x), we know the Python implementation is
incorrect for x, so the assertion fails.

This technique is very powerful, and is good at finding awkward edge
cases that people tend not to think of when writing test cases manually.
A large number of cases in tests/xlfunctions/test_engineering.py were
found by this method.